+++ src/Harmony/ViewModels/LibraryHubViewModel.cs
@@ -50,22 +50,20 @@ public partial class LibraryHubViewModel : ObservableObject
 
         LibraryShortcuts = new ObservableCollection<LibraryShortcutItem>
         {
-            new("library.songs", "library.subtitle", "F1 M4 10v2h12v-2H4zm0-4v2h12V6H4zm0 8v2h8v-2H4zm14-4v8.17c-.31-.11-.65-.17-1-.17-1.66 0-3 1.34-3 3s1.34 3 3 3 3-1.34 3-3V8h3V6h-5z", LibrarySection.Songs),
-            new("nav.playlists", "library.playlistsSubtitle", "F1 M15 6H3v2h12V6zm0 4H3v2h12v-2zM3 16h8v-2H3v2zM17 6v8.18c-.31-.11-.65-.18-1-.18-1.66 0-3 1.34-3 3s1.34 3 3 3 3-1.34 3-3V8h3V6h-5z", LibrarySection.Collections),
-            new("nav.favorites", "favorites.subtitle", "F1 M12 21.35l-1.45-1.32C5.4 15.36 2 12.28 2 8.5 2 5.42 4.42 3 7.5 3c1.74 0 3.41.81 4.5 2.09C13.09 3.81 14.76 3 16.5 3 19.58 3 22 5.42 22 8.5c0 3.78-3.4 6.86-8.55 11.54L12 21.35z", LibrarySection.Favorites),
-            new("nav.history", "library.historySubtitle", "F1 M13 3c-4.97 0-9 4.03-9 9H1l3.89 3.89.07.14L9 12H6c0-3.87 3.13-7 7-7s7 3.13 7 7-3.13 7-7 7c-1.93 0-3.68-.79-4.94-2.06l-1.42 1.42C8.27 19.99 10.51 21 13 21c4.97 0 9-4.03 9-9s-4.03-9-9-9zm-1 5v5l4.28 2.54.72-1.21-3.5-2.08V8H12z", LibrarySection.History),
+            new("library.songs", "F1 M4 10v2h12v-2H4zm0-4v2h12V6H4zm0 8v2h8v-2H4zm14-4v8.17c-.31-.11-.65-.17-1-.17-1.66 0-3 1.34-3 3s1.34 3 3 3 3-1.34 3-3V8h3V6h-5z", LibrarySection.Songs),
+            new("nav.playlists", "F1 M15 6H3v2h12V6zm0 4H3v2h12v-2zM3 16h8v-2H3v2zM17 6v8.18c-.31-.11-.65-.18-1-.18-1.66 0-3 1.34-3 3s1.34 3 3 3 3-1.34 3-3V8h3V6h-5z", LibrarySection.Collections),
+            new("nav.favorites", "F1 M12 21.35l-1.45-1.32C5.4 15.36 2 12.28 2 8.5 2 5.42 4.42 3 7.5 3c1.74 0 3.41.81 4.5 2.09C13.09 3.81 14.76 3 16.5 3 19.58 3 22 5.42 22 8.5c0 3.78-3.4 6.86-8.55 11.54L12 21.35z", LibrarySection.Favorites),
+            new("nav.history", "F1 M13 3c-4.97 0-9 4.03-9 9H1l3.89 3.89.07.14L9 12H6c0-3.87 3.13-7 7-7s7 3.13 7 7-3.13 7-7 7c-1.93 0-3.68-.79-4.94-2.06l-1.42 1.42C8.27 19.99 10.51 21 13 21c4.97 0 9-4.03 9-9s-4.03-9-9-9zm-1 5v5l4.28 2.54.72-1.21-3.5-2.08V8H12z", LibrarySection.History),
         };
 
         _loc.LanguageChanged += (_, _) => RefreshLabels();
         RefreshLabels();
-        SelectedSection = LibrarySection.Songs;
+        SelectedSection = LibrarySection.Collections;
         ShowSectionContent = true;
         ActivateSection(SelectedSection);
     }
 
     public ObservableCollection<LibraryShortcutItem> LibraryShortcuts { get; }
-    public string HubTitle => _loc.T("library.title");
-    public string HubSubtitle => _loc.T("library.hubSubtitle");
     public string SectionTitle => SelectedSection switch
     {
         LibrarySection.Songs => _loc.T("library.songs"),
@@ -99,7 +97,6 @@ public partial class LibraryHubViewModel : ObservableObject
 
         SelectedSection = page switch
         {
-            AppPage.Library => LibrarySection.Songs,
             AppPage.Collections or AppPage.Albums or AppPage.Playlists => LibrarySection.Collections,
             AppPage.Favorites => LibrarySection.Favorites,
             AppPage.History => LibrarySection.History,
@@ -170,13 +167,8 @@ public partial class LibraryHubViewModel : ObservableObject
     private void RefreshLabels()
     {
         foreach (var s in LibraryShortcuts)
-        {
             s.Title = _loc.T(s.TitleKey);
-            s.Subtitle = _loc.T(s.SubtitleKey);
-        }
         OnPropertyChanged(nameof(LibraryShortcuts));
-        OnPropertyChanged(nameof(HubTitle));
-        OnPropertyChanged(nameof(HubSubtitle));
         OnPropertyChanged(nameof(SectionTitle));
         OnPropertyChanged(nameof(SectionSubtitle));
     }